@@ -4,7 +4,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"log"
 	"regexp"
 	"strings"
@@ -142,19 +142,6 @@ func skipCtl(str string) []byte {
 		}
 	}
 	return []byte(str)
-}
-
-func stripCtl(str string) []byte {
-	b := make([]byte, len(str))
-	var bl int
-	for i := 0; i < len(str); i++ {
-		c := str[i]
-		if c >= 32 && c != 127 {
-			b[bl] = c
-			bl++
-		}
-	}
-	return b[:bl]
 }
 
 func (c *Ctl) CheckBackup(bcpName string, waitFor time.Duration) error {
@@ -357,7 +344,7 @@ func (c *Ctl) RunCmd(cmds ...string) (string, error) {
 				return "", errors.Wrap(err, "ContainerExecInspect")
 			}
 			if !insp.Running {
-				logs, err := ioutil.ReadAll(container.Reader)
+				logs, err := io.ReadAll(container.Reader)
 				if err != nil {
 					return "", errors.Wrap(err, "read logs of failed container")
 				}
@@ -383,7 +370,7 @@ func (c *Ctl) ContainerLogs() (string, error) {
 		return "", errors.Wrap(err, "get logs of failed container")
 	}
 	defer r.Close()
-	logs, err := ioutil.ReadAll(r)
+	logs, err := io.ReadAll(r)
 	if err != nil {
 		return "", errors.Wrap(err, "read logs of failed container")
 	}
