@@ -4,7 +4,6 @@ import (
 	"crypto/tls"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"net/url"
 	"path"
@@ -162,7 +161,7 @@ func (c *Conf) Cast() error {
 // Passing nil as the io.Writer will discard any warnings.
 func SDKLogLevel(levels string, out io.Writer) aws.LogLevelType {
 	if out == nil {
-		out = ioutil.Discard
+		out = io.Discard
 	}
 
 	var logLevel aws.LogLevelType
@@ -232,9 +231,13 @@ func New(opts Conf, l *log.Event) (*S3, error) {
 	return s, nil
 }
 
-const defaultPartSize = 10 * 1024 * 1024 // 10Mb
+const defaultPartSize int64 = 10 * 1024 * 1024 // 10Mb
 
-func (s *S3) Save(name string, data io.Reader, sizeb int) error {
+func (*S3) Type() storage.Type {
+	return storage.S3
+}
+
+func (s *S3) Save(name string, data io.Reader, sizeb int64) error {
 	switch s.opts.Provider {
 	default:
 		awsSession, err := s.session()
@@ -273,7 +276,7 @@ func (s *S3) Save(name string, data io.Reader, sizeb int) error {
 				s.opts.UploadPartSize = int(s3manager.MinUploadPartSize)
 			}
 
-			partSize = s.opts.UploadPartSize
+			partSize = int64(s.opts.UploadPartSize)
 		}
 		if sizeb > 0 {
 			ps := sizeb / s3manager.MaxUploadParts * 11 / 10 // add 10% just in case
@@ -282,15 +285,10 @@ func (s *S3) Save(name string, data io.Reader, sizeb int) error {
 			}
 		}
 
-		if s.log != nil {
-			s.log.Info("s3.uploadPartSize is set to %d (~%dMb)", partSize, partSize>>20)
-			s.log.Info("s3.maxUploadParts is set to %d", s.opts.MaxUploadParts)
-		}
-
 		_, err = s3manager.NewUploader(awsSession, func(u *s3manager.Uploader) {
 			u.MaxUploadParts = s.opts.MaxUploadParts
-			u.PartSize = int64(partSize) // 10MB part size
-			u.LeavePartsOnError = true   // Don't delete the parts if the upload fails.
+			u.PartSize = partSize      // 10MB part size
+			u.LeavePartsOnError = true // Don't delete the parts if the upload fails.
 			u.Concurrency = cc
 
 			u.RequestOptions = append(u.RequestOptions, func(r *request.Request) {
@@ -616,10 +614,15 @@ func (s *S3) session() (*session.Session, error) {
 		}})
 	}
 
+	ec2Session, err := session.NewSession()
+	if err != nil {
+		return nil, errors.WithMessage(err, "new session")
+	}
+
 	// allow fetching credentials from env variables and ec2 metadata endpoint
 	providers = append(providers, &credentials.EnvProvider{})
 	providers = append(providers, &ec2rolecreds.EC2RoleProvider{
-		Client: ec2metadata.New(session.New()),
+		Client: ec2metadata.New(ec2Session),
 	})
 
 	httpClient := http.DefaultClient
