@@ -1,18 +1,18 @@
 package speedt
 
 import (
-	"io/ioutil"
+	"io"
 	"testing"
 )
 
 func BenchmarkWriterBytes(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		ioutil.Discard.Write([]byte(dataset[i%len(dataset)]))
+		io.Discard.Write([]byte(dataset[i%len(dataset)]))
 	}
 }
 
 func BenchmarkWriterUnsafe(b *testing.B) {
 	for i := 0; i < b.N; i++ {
-		ioutil.Discard.Write(StringToBytes(dataset[i%len(dataset)]))
+		io.Discard.Write(StringToBytes(dataset[i%len(dataset)]))
 	}
 }
