@@ -115,11 +115,13 @@ func (a *Agent) Start() error {
 
 // Delete deletes backup(s) from the store and cleans up its metadata
 func (a *Agent) Delete(d *pbm.DeleteBackupCmd, opid pbm.OPID, ep pbm.Epoch) {
-	l := a.pbm.Logger().NewEvent(string(pbm.CmdDeleteBackup), "", opid.String(), ep.TS())
 	if d == nil {
+		l := a.log.NewEvent(string(pbm.CmdDeleteBackup), "", opid.String(), ep.TS())
 		l.Error("missed command")
 		return
 	}
+
+	l := a.pbm.Logger().NewEvent(string(pbm.CmdDeleteBackup), "", opid.String(), ep.TS())
 
 	nodeInfo, err := a.node.GetInfo()
 	if err != nil {
@@ -186,11 +188,13 @@ func (a *Agent) Delete(d *pbm.DeleteBackupCmd, opid pbm.OPID, ep pbm.Epoch) {
 
 // DeletePITR deletes PITR chunks from the store and cleans up its metadata
 func (a *Agent) DeletePITR(d *pbm.DeletePITRCmd, opid pbm.OPID, ep pbm.Epoch) {
-	l := a.pbm.Logger().NewEvent(string(pbm.CmdDeletePITR), "", opid.String(), ep.TS())
 	if d == nil {
+		l := a.log.NewEvent(string(pbm.CmdDeletePITR), "", opid.String(), ep.TS())
 		l.Error("missed command")
 		return
 	}
+
+	l := a.pbm.Logger().NewEvent(string(pbm.CmdDeletePITR), "", opid.String(), ep.TS())
 
 	nodeInfo, err := a.node.GetInfo()
 	if err != nil {
