@@ -260,11 +260,13 @@ func (a *Agent) pitrLockCheck() (moveOn bool, err error) {
 
 // PITRestore starts the point-in-time recovery
 func (a *Agent) PITRestore(r *pbm.PITRestoreCmd, opid pbm.OPID, ep pbm.Epoch) {
-	l := a.log.NewEvent(string(pbm.CmdPITRestore), time.Unix(r.TS, 0).UTC().Format(time.RFC3339), opid.String(), ep.TS())
 	if r == nil {
+		l := a.log.NewEvent(string(pbm.CmdPITRestore), "", opid.String(), ep.TS())
 		l.Error("missed command")
 		return
 	}
+
+	l := a.log.NewEvent(string(pbm.CmdPITRestore), time.Unix(r.TS, 0).UTC().Format(time.RFC3339), opid.String(), ep.TS())
 
 	nodeInfo, err := a.node.GetInfo()
 	if err != nil {
