@@ -8,11 +8,12 @@ import (
 	"time"
 
 	"github.com/alecthomas/kingpin"
-	"github.com/percona/percona-backup-mongodb/speedt"
 	"go.mongodb.org/mongo-driver/mongo"
 
 	"github.com/percona/percona-backup-mongodb/pbm"
+	"github.com/percona/percona-backup-mongodb/pbm/compress"
 	"github.com/percona/percona-backup-mongodb/pbm/storage/blackhole"
+	"github.com/percona/percona-backup-mongodb/speedt"
 	"github.com/percona/percona-backup-mongodb/version"
 )
 
@@ -27,11 +28,11 @@ func main() {
 		compressLevel    *int
 
 		compressType = tCmd.Flag("compression", "Compression type <none>/<gzip>/<snappy>/<lz4>/<s2>/<pgzip>/<zstd>").
-				Default(string(pbm.CompressionTypeS2)).
-				Enum(string(pbm.CompressionTypeNone), string(pbm.CompressionTypeGZIP),
-				string(pbm.CompressionTypeSNAPPY), string(pbm.CompressionTypeLZ4),
-				string(pbm.CompressionTypeS2), string(pbm.CompressionTypePGZIP),
-				string(pbm.CompressionTypeZstandard),
+				Default(string(compress.CompressionTypeS2)).
+				Enum(string(compress.CompressionTypeNone), string(compress.CompressionTypeGZIP),
+				string(compress.CompressionTypeSNAPPY), string(compress.CompressionTypeLZ4),
+				string(compress.CompressionTypeS2), string(compress.CompressionTypePGZIP),
+				string(compress.CompressionTypeZstandard),
 			)
 
 		compressionCmd = tCmd.Command("compression", "Run compression test")
@@ -62,10 +63,10 @@ func main() {
 	switch cmd {
 	case compressionCmd.FullCommand():
 		fmt.Print("Test started ")
-		compression(*mURL, pbm.CompressionType(*compressType), compressLevel, *sampleSizeF, *sampleColF)
+		compression(*mURL, compress.CompressionType(*compressType), compressLevel, *sampleSizeF, *sampleColF)
 	case storageCmd.FullCommand():
 		fmt.Print("Test started ")
-		storage(*mURL, pbm.CompressionType(*compressType), compressLevel, *sampleSizeF, *sampleColF)
+		storage(*mURL, compress.CompressionType(*compressType), compressLevel, *sampleSizeF, *sampleColF)
 	case versionCmd.FullCommand():
 		switch {
 		case *versionCommit:
@@ -78,7 +79,7 @@ func main() {
 	}
 }
 
-func compression(mURL string, compression pbm.CompressionType, level *int, sizeGb float64, collection string) {
+func compression(mURL string, compression compress.CompressionType, level *int, sizeGb float64, collection string) {
 	ctx := context.Background()
 
 	var cn *mongo.Client
@@ -88,7 +89,7 @@ func compression(mURL string, compression pbm.CompressionType, level *int, sizeG
 		if err != nil {
 			log.Fatalln("Error: connect to mongodb-node:", err)
 		}
-		defer node.Session().Disconnect(ctx)
+		defer func() { node.Session().Disconnect(ctx) }()
 
 		cn = node.Session()
 	}
@@ -107,20 +108,20 @@ func compression(mURL string, compression pbm.CompressionType, level *int, sizeG
 	fmt.Println(r)
 }
 
-func storage(mURL string, compression pbm.CompressionType, level *int, sizeGb float64, collection string) {
+func storage(mURL string, compression compress.CompressionType, level *int, sizeGb float64, collection string) {
 	ctx := context.Background()
 
 	node, err := pbm.NewNode(ctx, mURL, 1)
 	if err != nil {
 		log.Fatalln("Error: connect to mongodb-node:", err)
 	}
-	defer node.Session().Disconnect(ctx)
+	defer func() { node.Session().Disconnect(ctx) }()
 
 	pbmClient, err := pbm.New(ctx, mURL, "pbm-speed-test")
 	if err != nil {
 		log.Fatalln("Error: connect to mongodb-pbm:", err)
 	}
-	defer pbmClient.Conn.Disconnect(ctx)
+	defer func() { pbmClient.Conn.Disconnect(ctx) }()
 
 	stg, err := pbmClient.GetStorage(nil)
 	if err != nil {
