@@ -16,16 +16,6 @@ import (
 	"github.com/percona/percona-backup-mongodb/e2e-tests/pkg/pbm"
 )
 
-type shard struct {
-	name string
-	cn   *pbm.Mongo
-}
-
-type trxData struct {
-	Country string
-	UID     int
-}
-
 const trxdb = "trx"
 
 func (c *Cluster) DistributedTransactions(bcp Backuper, col string) {
@@ -116,7 +106,7 @@ func (c *Cluster) DistributedTransactions(bcp Backuper, col string) {
 	// distributed transaction that commits before the backup ends
 	// should be visible after restore
 	log.Println("Run trx1")
-	_, _ = sess.WithTransaction(ctx, func(sc mongo.SessionContext) (interface{}, error) {
+	sess.WithTransaction(ctx, func(sc mongo.SessionContext) (interface{}, error) {
 		c.trxSet(sc, 30, col)
 		c.trxSet(sc, 530, col)
 
@@ -140,14 +130,14 @@ func (c *Cluster) DistributedTransactions(bcp Backuper, col string) {
 	log.Println("Run trx2")
 	// distributed transaction that commits after the backup ends
 	// should NOT be visible after the restore
-	_ = mongo.WithSession(ctx, sess, func(sc mongo.SessionContext) error {
+	mongo.WithSession(ctx, sess, func(sc mongo.SessionContext) error {
 		err := sess.StartTransaction()
 		if err != nil {
 			log.Fatalln("ERROR: start transaction:", err)
 		}
 		defer func() {
 			if err != nil {
-				_ = sess.AbortTransaction(sc)
+				sess.AbortTransaction(sc)
 				log.Fatalln("ERROR: transaction:", err)
 			}
 		}()
