@@ -5,14 +5,22 @@ import (
 	"context"
 	"encoding/json"
 	"io"
+	"path"
+	"strings"
+	"sync/atomic"
 	"time"
 
 	mlog "github.com/mongodb/mongo-tools/common/log"
 	"github.com/pkg/errors"
+	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+	"golang.org/x/sync/errgroup"
 
 	"github.com/percona/percona-backup-mongodb/pbm"
+	"github.com/percona/percona-backup-mongodb/pbm/archive"
+	"github.com/percona/percona-backup-mongodb/pbm/compress"
 	plog "github.com/percona/percona-backup-mongodb/pbm/log"
 	"github.com/percona/percona-backup-mongodb/pbm/storage"
 	"github.com/percona/percona-backup-mongodb/version"
@@ -58,6 +66,7 @@ func (b *Backup) Init(bcp *pbm.BackupCmd, opid pbm.OPID, balancer pbm.BalancerMo
 		Type:           b.typ,
 		OPID:           opid.String(),
 		Name:           bcp.Name,
+		Namespaces:     bcp.Namespaces,
 		Compression:    bcp.Compression,
 		StartTS:        time.Now().Unix(),
 		Status:         pbm.StatusStarting,
@@ -101,6 +110,9 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 		Conditions:   []pbm.Condition{},
 		FirstWriteTS: primitive.Timestamp{T: 1, I: 1},
 	}
+	if v := inf.IsConfigSrv(); v {
+		rsMeta.IsConfigSvr = &v
+	}
 
 	stg, err := b.cn.GetStorage(l)
 	if err != nil {
@@ -118,14 +130,6 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 			status := pbm.StatusError
 			if errors.Is(err, ErrCancelled) {
 				status = pbm.StatusCancelled
-
-				meta := &pbm.BackupMeta{
-					Name:     bcp.Name,
-					Status:   pbm.StatusCancelled,
-					Replsets: []pbm.BackupReplset{rsMeta},
-				}
-
-				l.Info("delete artifacts from storage: %v", b.cn.DeleteBackupFiles(meta, stg))
 			}
 
 			ferr := b.cn.ChangeRSState(bcp.Name, rsMeta.Name, status, err.Error())
@@ -178,6 +182,13 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 			}
 		}()
 
+		if inf.IsConfigSrv() && len(bcp.Namespaces) != 0 {
+			err := checkNamespaceForBackup(ctx, b.cn.Conn, bcp.Namespaces[0])
+			if err != nil {
+				return errors.WithMessage(err, "namespace check")
+			}
+		}
+
 		if bcpm.BalancerStatus == pbm.BalancerModeOn {
 			err = b.cn.SetBalancerStatus(pbm.BalancerModeOff)
 			if err != nil {
@@ -195,6 +206,16 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 	if err != nil {
 		return errors.Wrap(err, "waiting for start")
 	}
+
+	defer func() {
+		if !errors.Is(err, ErrCancelled) || !inf.IsLeader() {
+			return
+		}
+
+		if err := b.cn.DeleteBackupFiles(bcpm, stg); err != nil {
+			l.Error("Failed to delete leftover files for canceled backup %q", bcpm.Name)
+		}
+	}()
 
 	switch b.typ {
 	case pbm.LogicalBackup:
@@ -226,7 +247,29 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 			return errors.Wrap(err, "check cluster for backup done")
 		}
 
-		err = b.dumpClusterMeta(bcp.Name, stg)
+		bcpm, err = b.cn.GetBackupMeta(bcp.Name)
+		if err != nil {
+			return errors.Wrap(err, "get backup metadata")
+		}
+
+		var backupSize int64
+		switch b.typ {
+		case pbm.LogicalBackup:
+			backupSize, err = countLogicalBackupSize(ctx, bcpm, stg)
+			if err != nil {
+				return errors.Wrap(err, "count backup size")
+			}
+		case pbm.PhysicalBackup:
+			backupSize = countPhysicalBackupSize(ctx, bcpm, stg)
+		}
+
+		err = b.cn.SetBackupSize(ctx, bcp.Name, backupSize)
+		if err != nil {
+			return errors.Wrap(err, "set backup size")
+		}
+		bcpm.Size = backupSize
+
+		err = writeMeta(stg, bcpm)
 		if err != nil {
 			return errors.Wrap(err, "dump metadata")
 		}
@@ -235,6 +278,121 @@ func (b *Backup) Run(ctx context.Context, bcp *pbm.BackupCmd, opid pbm.OPID, l *
 	// to be sure the locks released only after the "done" status had written
 	err = b.waitForStatus(bcp.Name, pbm.StatusDone, nil)
 	return errors.Wrap(err, "waiting for done")
+}
+
+func countPhysicalBackupSize(ctx context.Context, bcp *pbm.BackupMeta, stg storage.Storage) int64 {
+	totalSize := int64(0)
+
+	for _, rs := range bcp.Replsets {
+		for _, f := range rs.Files {
+			totalSize += f.StgSize
+		}
+	}
+
+	return totalSize
+}
+
+func countLogicalBackupSize(ctx context.Context, bcp *pbm.BackupMeta, stg storage.Storage) (int64, error) {
+	totalSize := int64(0)
+
+	eg, _ := errgroup.WithContext(ctx)
+	for _, rs := range bcp.Replsets {
+		rs := &rs
+
+		eg.Go(func() error {
+			f, err := stg.FileStat(rs.DumpName)
+			if err != nil {
+				return errors.WithMessagef(err, "dump file %q", f)
+			}
+			atomic.AddInt64(&totalSize, f.Size)
+			return nil
+		})
+
+		eg.Go(func() error {
+			f, err := stg.FileStat(rs.OplogName)
+			if err != nil {
+				return errors.WithMessagef(err, "oplog file %q", f)
+			}
+			atomic.AddInt64(&totalSize, f.Size)
+			return nil
+		})
+
+		metafile := path.Join(bcp.Name, rs.Name, archive.MetaFile)
+		nss, err := pbm.ReadArchiveNamespaces(stg, metafile)
+		if err != nil {
+			return 0, errors.WithMessage(err, "read archive namespaces")
+		}
+
+		for _, ns := range nss {
+			if ns.Size == 0 {
+				continue
+			}
+
+			ns := archive.NSify(ns.Database, ns.Collection)
+			filename := path.Join(bcp.Name, rs.Name, ns+bcp.Compression.Suffix())
+
+			eg.Go(func() error {
+				f, err := stg.FileStat(filename)
+				if err != nil {
+					return errors.WithMessagef(err, "namespace %q file %q", ns, filename)
+				}
+				atomic.AddInt64(&totalSize, f.Size)
+				return nil
+			})
+		}
+	}
+
+	err := eg.Wait()
+	return atomic.LoadInt64(&totalSize), err
+}
+
+func checkNamespaceForBackup(ctx context.Context, m *mongo.Client, ns string) error {
+	db, coll := parseNS(ns)
+
+	res, err := getShardedNamespaces(ctx, m, db, coll)
+	if err != nil {
+		return errors.WithMessage(err, "get sharded namespaces")
+	}
+	if len(res) != 0 {
+		return errors.Errorf("selective backup: sharded collections: %s", strings.Join(res, ", "))
+	}
+
+	return nil
+}
+
+func getShardedNamespaces(ctx context.Context, m *mongo.Client, db, coll string) ([]string, error) {
+	cur, err := m.Database("config").Collection("collections").
+		Find(ctx, bson.D{}, options.Find().SetProjection(bson.D{{"_id", 1}}))
+	if err != nil {
+		return nil, errors.WithMessage(err, "query")
+	}
+	defer cur.Close(context.Background())
+
+	type document struct {
+		ID string `bson:"_id"`
+	}
+
+	rv := []string{}
+	for cur.Next(ctx) {
+		var doc document
+		if err := cur.Decode(&doc); err != nil {
+			return nil, errors.WithMessage(err, "decode")
+		}
+
+		d, c, _ := strings.Cut(doc.ID, ".")
+		if db != "" && db != d {
+			continue
+		}
+
+		c = strings.TrimPrefix(c, "system.buckets.")
+		if coll != "" && coll != c {
+			continue
+		}
+
+		rv = append(rv, d+"."+c)
+	}
+
+	return rv, nil
 }
 
 func waitForBalancerOff(cn *pbm.PBM, t time.Duration, l *plog.Event) pbm.BalancerMode {
@@ -325,10 +483,10 @@ type Canceller interface {
 var ErrCancelled = errors.New("backup canceled")
 
 // Upload writes data to dst from given src and returns an amount of written bytes
-func Upload(ctx context.Context, src Source, dst storage.Storage, compression pbm.CompressionType, compressLevel *int, fname string, sizeb int) (int64, error) {
+func Upload(ctx context.Context, src Source, dst storage.Storage, compression compress.CompressionType, compressLevel *int, fname string, sizeb int64) (int64, error) {
 	r, pw := io.Pipe()
 
-	w, err := Compress(pw, compression, compressLevel)
+	w, err := compress.Compress(pw, compression, compressLevel)
 	if err != nil {
 		return 0, err
 	}
@@ -557,15 +715,6 @@ func (b *Backup) waitForFirstLastWrite(bcpName string) (first, last primitive.Ti
 			return first, last, nil
 		}
 	}
-}
-
-func (b *Backup) dumpClusterMeta(bcpName string, stg storage.Storage) error {
-	meta, err := b.cn.GetBackupMeta(bcpName)
-	if err != nil {
-		return errors.Wrap(err, "get backup metadata")
-	}
-
-	return writeMeta(stg, meta)
 }
 
 func writeMeta(stg storage.Storage, meta *pbm.BackupMeta) error {
