@@ -15,6 +15,7 @@ import (
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"gopkg.in/yaml.v2"
 
+	"github.com/percona/percona-backup-mongodb/pbm/compress"
 	"github.com/percona/percona-backup-mongodb/pbm/log"
 	"github.com/percona/percona-backup-mongodb/pbm/storage"
 	"github.com/percona/percona-backup-mongodb/pbm/storage/azure"
@@ -59,49 +60,38 @@ func (c Config) String() string {
 
 // PITRConf is a Point-In-Time Recovery options
 type PITRConf struct {
-	Enabled          bool            `bson:"enabled" json:"enabled" yaml:"enabled"`
-	OplogSpanMin     float64         `bson:"oplogSpanMin" json:"oplogSpanMin" yaml:"oplogSpanMin"`
-	OplogOnly        bool            `bson:"oplogOnly,omitempty" json:"oplogOnly,omitempty" yaml:"oplogOnly,omitempty"`
-	Compression      CompressionType `bson:"compression,omitempty" json:"compression,omitempty" yaml:"compression,omitempty"`
-	CompressionLevel *int            `bson:"compressionLevel,omitempty" json:"compressionLevel,omitempty" yaml:"compressionLevel,omitempty"`
+	Enabled          bool                     `bson:"enabled" json:"enabled" yaml:"enabled"`
+	OplogSpanMin     float64                  `bson:"oplogSpanMin" json:"oplogSpanMin" yaml:"oplogSpanMin"`
+	OplogOnly        bool                     `bson:"oplogOnly,omitempty" json:"oplogOnly,omitempty" yaml:"oplogOnly,omitempty"`
+	Compression      compress.CompressionType `bson:"compression,omitempty" json:"compression,omitempty" yaml:"compression,omitempty"`
+	CompressionLevel *int                     `bson:"compressionLevel,omitempty" json:"compressionLevel,omitempty" yaml:"compressionLevel,omitempty"`
 }
 
 func (c *PITRConf) Cast() error {
 	if c.Compression == "" {
-		c.Compression = CompressionTypeS2
+		c.Compression = compress.CompressionTypeS2
 	}
 
 	return nil
 }
 
-// StorageType represents a type of the destination storage for backups
-type StorageType string
-
-const (
-	StorageUndef      StorageType = ""
-	StorageS3         StorageType = "s3"
-	StorageAzure      StorageType = "azure"
-	StorageFilesystem StorageType = "filesystem"
-	StorageBlackHole  StorageType = "blackhole"
-)
-
 // StorageConf is a configuration of the backup storage
 type StorageConf struct {
-	Type       StorageType `bson:"type" json:"type" yaml:"type"`
-	S3         s3.Conf     `bson:"s3,omitempty" json:"s3,omitempty" yaml:"s3,omitempty"`
-	Azure      azure.Conf  `bson:"azure,omitempty" json:"azure,omitempty" yaml:"azure,omitempty"`
-	Filesystem fs.Conf     `bson:"filesystem,omitempty" json:"filesystem,omitempty" yaml:"filesystem,omitempty"`
+	Type       storage.Type `bson:"type" json:"type" yaml:"type"`
+	S3         s3.Conf      `bson:"s3,omitempty" json:"s3,omitempty" yaml:"s3,omitempty"`
+	Azure      azure.Conf   `bson:"azure,omitempty" json:"azure,omitempty" yaml:"azure,omitempty"`
+	Filesystem fs.Conf      `bson:"filesystem,omitempty" json:"filesystem,omitempty" yaml:"filesystem,omitempty"`
 }
 
 func (s *StorageConf) Typ() string {
 	switch s.Type {
-	case StorageS3:
+	case storage.S3:
 		return "S3"
-	case StorageAzure:
+	case storage.Azure:
 		return "Azure"
-	case StorageFilesystem:
+	case storage.Filesystem:
 		return "FS"
-	case StorageBlackHole:
+	case storage.BlackHole:
 		return "BlackHole"
 	default:
 		return "Unknown"
@@ -111,7 +101,7 @@ func (s *StorageConf) Typ() string {
 func (s *StorageConf) Path() string {
 	path := ""
 	switch s.Type {
-	case StorageS3:
+	case storage.S3:
 		path = "s3://"
 		if s.S3.EndpointURL != "" {
 			path += s.S3.EndpointURL + "/"
@@ -120,14 +110,14 @@ func (s *StorageConf) Path() string {
 		if s.S3.Prefix != "" {
 			path += "/" + s.S3.Prefix
 		}
-	case StorageAzure:
+	case storage.Azure:
 		path = fmt.Sprintf(azure.BlobURL, s.Azure.Account, s.Azure.Container)
 		if s.Azure.Prefix != "" {
 			path += "/" + s.Azure.Prefix
 		}
-	case StorageFilesystem:
+	case storage.Filesystem:
 		path = s.Filesystem.Path
-	case StorageBlackHole:
+	case storage.BlackHole:
 		path = "BlackHole"
 	}
 
@@ -141,9 +131,9 @@ type RestoreConf struct {
 }
 
 type BackupConf struct {
-	Priority         map[string]float64 `bson:"priority,omitempty" json:"priority,omitempty" yaml:"priority,omitempty"`
-	Compression      CompressionType    `bson:"compression,omitempty" json:"compression,omitempty" yaml:"compression,omitempty"`
-	CompressionLevel *int               `bson:"compressionLevel,omitempty" json:"compressionLevel,omitempty" yaml:"compressionLevel,omitempty"`
+	Priority         map[string]float64       `bson:"priority,omitempty" json:"priority,omitempty" yaml:"priority,omitempty"`
+	Compression      compress.CompressionType `bson:"compression,omitempty" json:"compression,omitempty" yaml:"compression,omitempty"`
+	CompressionLevel *int                     `bson:"compressionLevel,omitempty" json:"compressionLevel,omitempty" yaml:"compressionLevel,omitempty"`
 }
 
 type confMap map[string]reflect.Kind
@@ -187,7 +177,7 @@ func (p *PBM) SetConfigByte(buf []byte) error {
 
 func (p *PBM) SetConfig(cfg Config) error {
 	switch cfg.Storage.Type {
-	case StorageS3:
+	case storage.S3:
 		err := cfg.Storage.S3.Cast()
 		if err != nil {
 			return errors.Wrap(err, "cast storage")
@@ -196,14 +186,14 @@ func (p *PBM) SetConfig(cfg Config) error {
 		// call the function for notification purpose.
 		// warning about unsupported levels will be printed
 		s3.SDKLogLevel(cfg.Storage.S3.DebugLogLevels, os.Stderr)
-	case StorageFilesystem:
+	case storage.Filesystem:
 		err := cfg.Storage.Filesystem.Cast()
 		if err != nil {
 			return errors.Wrap(err, "check config")
 		}
 	}
 
-	if c := string(cfg.PITR.Compression); c != "" && !isValidCompressionType(c) {
+	if c := string(cfg.PITR.Compression); c != "" && !compress.IsValidCompressionType(c) {
 		return errors.Errorf("unsupported compression type: %q", c)
 	}
 
@@ -261,7 +251,7 @@ func (p *PBM) SetConfigVar(key, val string) error {
 	case "pitr.enabled":
 		return errors.Wrap(p.confSetPITR(key, v.(bool)), "write to db")
 	case "pitr.compression":
-		if c := v.(string); c != "" && !isValidCompressionType(c) {
+		if c := v.(string); c != "" && !compress.IsValidCompressionType(c) {
 			return errors.Errorf("unsupported compression type: %q", c)
 		}
 	case "storage.filesystem.path":
@@ -415,15 +405,15 @@ func (p *PBM) GetStorage(l *log.Event) (storage.Storage, error) {
 // Storage creates and returns a storage object based on a given config
 func Storage(c Config, l *log.Event) (storage.Storage, error) {
 	switch c.Storage.Type {
-	case StorageS3:
+	case storage.S3:
 		return s3.New(c.Storage.S3, l)
-	case StorageAzure:
+	case storage.Azure:
 		return azure.New(c.Storage.Azure, l)
-	case StorageFilesystem:
+	case storage.Filesystem:
 		return fs.New(c.Storage.Filesystem), nil
-	case StorageBlackHole:
+	case storage.BlackHole:
 		return blackhole.New(), nil
-	case StorageUndef:
+	case storage.Undef:
 		return nil, ErrStorageUndefined
 	default:
 		return nil, errors.Errorf("unknown storage type %s", c.Storage.Type)
