@@ -9,7 +9,7 @@ import (
 )
 
 // current PBM version
-const version = "1.8.1"
+const version = "2.0.0-dev"
 
 // !!! should be sorted in the ascending order
 var breakingChangesV = []string{
@@ -140,4 +140,8 @@ func majmin(v string) string {
 	}
 
 	return semver.MajorMinor(v)
+}
+
+func IsLegacyArchive(ver string) bool {
+	return semver.Compare(majmin(ver), "v2.0") == -1
 }
